Module Brr.Abort

Aborting browser activities.

This mecanism provides a way to abort futures with a Jv.Error.t named AbortError (`Abort_error in Jv.Error.enum).

module Signal : sig ... end

Abort signals.

type t

The type for AbortController objects.

val controller : unit -> t

controller () is a new abort controller.

val signal : t -> Signal.t

signal c is the signal associated to abort controller c.

val abort : t -> unit

abort c aborts the signal of c and informs its observers the associated activity is to be aborted.