Module Abort.Signal

Abort signals.

type t

The type for AbortSignal objects.

val as_target : t -> Ev.target

as_target s is the signal as an event target.

val aborted : t -> bool

aborted s is true if s has signaled to abort.

val abort : Ev.void

abort is the type for abort events.