Module Brr.At

DOM element attributes.

Attributes

type name = Jstr.t

The type for attribute names.

type t

The type for attributes.

val v : name -> Jstr.t -> t

v n value is an attribute named n with value value.

val void : t

void is an attribute that doesn't exist. It is ignored by functions like El.v. This is v Jstr.empty Jstr.empty.

val is_void : t -> bool

is_void a is true iff a is void.

val true' : name -> t

true' n is v n Jstr.empty. This sets the boolean attribute n to true. The attribute must be omitted to be false.

val int : name -> int -> t

int n i is v n (Jstr.of_int i).

val float : name -> float -> t

float n f is v n (Jstr.of_float f).

val if' : bool -> t -> t

if' b a is a if b is true and void otherwise.

val if_some : t option -> t

if_some o is a if o is Some a and void if o is None.

val to_pair : t -> Jstr.t * Jstr.t

to_pair at is (n,v) the name and value of the attribute.

val add_if : bool -> t -> t list -> t list

add_if c att atts is att :: atts if c is true and atts otherwise.

deprecated use Brr.At.if' instead.

val add_if_some : name -> Jstr.t option -> t list -> t list

add_if_some n o atts is (v n value) :: atts if o is Some value and atts otherwise.

deprecated use Brr.At.if_some instead.

Attribute names and constructors

See the MDN HTML attribute reference.

Convention. Whenever an attribute name conflicts with an OCaml keyword we prime it, see for example class'.

module Name : sig ... end

Attribute names.

type 'a cons = 'a -> t

The type for attribute constructors with value of type 'a.

val accesskey : Jstr.t cons

accesskey

val action : Jstr.t cons

action

val autocomplete : Jstr.t cons

autocomplete

val autofocus : t

autofocus

val charset : Jstr.t cons

charset

val checked : t

checked

val class' : Jstr.t cons

class

val cols : int cons

cols

val content : Jstr.t cons

content

val contenteditable : bool cons

contenteditable

val defer : t

defer

val dir : Jstr.t cons

dir

val disabled : t

disabled

val download : Jstr.t cons

download

val draggable : Jstr.t cons

draggable

val for' : Jstr.t cons

for'

val height : int cons

height

val hidden : t

hidden

val href : Jstr.t cons

href

val id : Jstr.t cons

id

val lang : Jstr.t cons

lang

val list : Jstr.t cons

list

val media : Jstr.t cons

media

val method' : Jstr.t cons

method.

val name : Jstr.t cons

name

val placeholder : Jstr.t cons

placeholder

val rel : Jstr.t cons

rel

val required : t

required

val rows : int cons

rows

val selected : t

selected

val spellcheck : Jstr.t cons

spellcheck

val src : Jstr.t cons

src

val style : Jstr.t cons

style

val tabindex : int cons

tabindex

val title : Jstr.t cons

title

val type' : Jstr.t cons

type

val value : Jstr.t cons

value

val wrap : Jstr.t cons

wrap

val width : int cons

width