Module El.Input

The HTML input element interface .

val files : t -> File.t list

files e is the file list held by element e. Usually e is an El.input of At.type' file. The empty list is returned if e has no such list.