Module Ev.Type

Event types.

type void

A type for events that do not expose further data.

type 'a t = 'a type'

See type'.

val create : Jstr.t -> 'a type'

create n is a new event type named n. Constrain the result to your event type. See the predefined types.

val void : Jstr.t -> void t

void is a new void event type.

val name : 'a type' -> Jstr.t

name t is the name of event type t.