Module Ev.Wheel

Wheel events.

module Delta_mode : sig ... end

Delta unit specification.

type t

The type for WheelEvent objects.

val as_mouse_event : t -> Mouse.t event

as_mouse_event w is w as a mouse event.

val delta_x : t -> float

delta_x w is the amount to be scrolled on the x-axis.

val delta_y : t -> float

delta_x w is the amount to be scrolled on the y-axis.

val delta_z : t -> float

delta_z w is the amount to be scrolled on the z-axis.

val delta_mode : t -> Delta_mode.t

delta_mode w is the unit of measurement for delta_x, delta_y and delta_z.