Module Brr.Navigator

Navigator objects.

Some of the objects that are accessed from Navigator objects have the accessor in the module that handles them:

Brr_io.Clipboard.of_navigator

Brr_io.Geolocation.of_navigator

Brr_io.Media.Devices.of_navigator

Brr_webmidi.Midi.Access.of_navigator

type t

The type for navigator Navigator objects. See G.navigator for the global navigator object.

val languages : t -> Jstr.t list

languages n is the user's preferred languages as BCP 47 language tags ordered by decreasing preference.

This consults navigator.languages and navigator.language. See also Ev.languagechange

val max_touch_points : t -> int

max_touch_points n is the maximum number of simultaneous touch contacts supported by the user agent. See the pointer events spec.

val online : t -> bool

online n is the online status of the browser. See the docs, the semantics is browser dependent.