Module Brr.Performance

Performance objects.

See the Peformance API and Peformance Timeline API.

module Entry : sig ... end

Performance entry objects.

type t

The type for performance Performance objects. See G.performance for the global performance object.

val time_origin_ms : t -> float

time_origin_ms p is the start time of the performance measurement.

val clear_marks : t -> Jstr.t option -> unit

clear_marks p n clears the marks named n or all of them on None.

val clear_measures : t -> Jstr.t option -> unit

clear_measures p n clears the measures named n or all of them on None.

val clear_resource_timings : t -> unit

clear_measures p n clears the resource.

val get_entries : ?type':Entry.Type.t -> ?name:Jstr.t -> t -> Entry.t list

get_entries ~type' ~name' p are p's entries in chronological order filtered by given type' and/or name' (both can be omited, possibly separately).

val mark : t -> Jstr.t -> unit

mark p name creates an timestamped entry associated to name name.

val measure : ?start:Jstr.t -> ?stop:Jstr.t -> t -> Jstr.t -> unit

measure p n ~start ~stop creates an entry to measure time between two marks.

val now_ms : t -> float

now_ms p is the number of millisecond elapsed since time_origin_ms.

val to_json : t -> Json.t

to_json p is p converted to a JSON object.