Module Fetch.Cache

Fetch caches.

type query_opts

The type for query options.

val query_opts : ?ignore_search:bool -> ?ignore_method:bool -> ?ignore_vary:bool -> ?cache_name:Jstr.t -> unit -> query_opts

query_opts ~ignore_search ~ignore_method ~ignore_vary ~cache_name () are query options with given parameters.

type t

The type for Cache objects.

val match' : ?query_opts:query_opts -> t -> Request.t -> Response.t option Fut.or_error

match' c req is a stored response for req in c (if any).

val match_all : ?query_opts:query_opts -> t -> Request.t -> Response.t list Fut.or_error

match_all c req is a list stored response for req in c.

val add : t -> Request.t -> unit Fut.or_error

add c req fetches req and adds the response to c.

val add_all : t -> Request.t list -> unit Fut.or_error

add_all c reqs fetches reqs and adds their reponses to c.

val put : t -> Request.t -> Response.t -> unit Fut.or_error

put c req resp puts the req/resp pair to the cache.

val delete : ?query_opts:query_opts -> t -> Request.t -> bool Fut.or_error

delete c req deletes response to req from the cache. false is returned if req was not in the cache.

val keys : ?query_opts:query_opts -> ?req:Request.t -> t -> Request.t list Fut.or_error

keys c are the requests cached by c.

Cache storage

module Storage : sig ... end

Cache storage objects.