Module Geolocation.Pos

Positions.

type t

The type for GeolocationPosition objects (and their GeolocationCoordinates member).

val latitude : t -> float

latitude p is the latitude in decimal degrees.

val longitude : t -> float

longitude p is the longitude in decimal degrees.

val accuracy : t -> float

accuracy p is the accuracy, in meters, of the latitude and longitude in meters.

val altitude : t -> float option

altitude p is the altitude in meters relative to sea level.

val altitude_accuracy : t -> float option

altitude_accuracy p is the altitude accuracy, in meters, of the altitude.

val heading : t -> float option

heading p is the direction in degree with respect to true north (90° is east). If speed is 0, this is nan.

val speed : t -> float option

speed p is the device velocity in meters per seconds.

val timestamp_ms : t -> float

timestamp_ms p is the time of measurement in ms since the epoch.