Module Media.Constraints

Property constraints specifications.

type t

The type for constraints.

val empty : unit -> t

empty () is an empty set of constraints.

val find : ('a, 'b, 'c) Prop.t -> t -> 'c option

find p s is the constraint for p in c (if any).

val set : ('a, 'b, 'c) Prop.t -> 'c -> t -> unit

set p v c sets the constraint for p to v in c.

val delete : ('a, 'b, 'c) Prop.t -> t -> unit

delete p c deletes the constraint for p from c.