Module Device.Info

Device information.

type t

The type for MediaDeviceInfo objects.

val device_id : t -> Jstr.t

device_id d is the identifier of the device.

val kind : t -> Kind.t

kind d is the kind of device.

val label : t -> Jstr.t

label d is a label describing the device.

val group_id : t -> Jstr.t

group_id d is the group identifier of the device. Two devices have the same group identifier if they belong to the same physical device.

val to_json : t -> Brr.Json.t

to_json d is d as JSON data.