Module El.Time_ranges

Time ranges.

type t

The type for TimeRange objects.

val length : t -> int

length r is the length of r.

val start : t -> int -> float

start r i is the start time of range i in r.

val end' : t -> int -> float

end' r i is the end time of range i in r.