Module Ev.Blob

Blob events.

type t

The type for BlobEvent objects.

val data : t -> Brr.Blob.t

data e is the requested data as a blob object.

val timecode : t -> float

timecode e is the difference between timestamp of the first chunk in data and the one produced by the first chunk in the first blob event produced by the recorder (that one may not be zero).