Module Media.Settings

Property values.

type t

The type for settings.

val get : ('a, 'b, 'c) Prop.t -> t -> 'a

get p s is the value of p in s.

val find : ('a, 'b, 'c) Prop.t -> t -> 'a option

find p s is the value of p in s (if any).