Module Media.Supported_constraints

Supported property constraints.

Indicates the media properties constraints the user agent understands.

type t

The type for supported constraints.

val mem : ('a, 'b, 'c) Prop.t -> t -> bool

supports p n is true if property p can be constrained.

val names : t -> Jstr.t list

supported s is the list of supported constraints.