Module Brr_io.Message

Message events, ports, channels and broadcast channels.

type transfer

The type for objects to transfer.

val transfer : 'a -> transfer

transfer v indicates valule v should be transfered, not just cloned, meaning they are no longer usable on the sending side.

type opts

The type for messaging options.

val opts : ?target_origin:Jstr.t -> ?transfer:transfer list -> unit -> opts

opts ~target_origin ~transfer () are messaging options. See here for the semantics of target_origin and transfer.

module Port : sig ... end

Message ports.

module Channel : sig ... end

Message channels.

module Broadcast_channel : sig ... end

Broadcast channels.

val window_post : ?opts:opts -> Brr.Window.t -> 'a -> unit

window_post w v ~opts posts value v to window w with options opts.

Events

module Ev : sig ... end

Message events.