Module Brr_io.Storage

Storage objects.

See Web Storage API

type t

The type for Storage objects.

val local : Brr.Window.t -> t

local w is the storage saved accross page sessions for the window's origin.

val session : Brr.Window.t -> t

session w is the storage cleared when the page session ends for the window's origin.

val length : t -> int

length s is the number of items in s.

val key : t -> int -> Jstr.t option

key s i is the name of the ith key. (N.B. local storage can race with other tabs)

val get_item : t -> Jstr.t -> Jstr.t option

get_item s k is the value of k in s.

val set_item : t -> Jstr.t -> Jstr.t -> (unit, Jv.Error.t) result

set_item s k v sets the value of k to v in s. An error is returned if the value could not be set (no permission or quota exceeded).

val remove_item : t -> Jstr.t -> unit

remove_item s k removes the value of k from s. If k has no value this does nothing.

val clear : t -> unit

clear s removes all keys from s.

Events

module Ev : sig ... end

Storage event.