Module Consoler.S

Logging signal changes.

val log : ?obs:bool -> ?l:Brr.Console.log -> ?v:'a Brr.Console.msgr -> string -> 'a Note.signal -> 'a Note.signal

signal is like log_value but on the changes of s. If obs is true, the return value is s itself and the tracing occurs through a logger, note that this prevents s from being garbage collected. If obs is false, the return value is s mapped by a tracing identity and using s's equality function.