Module Brr_note.Futr

Futures as Note events and vice-versa.

val to_event : 'a Fut.t -> 'a Note.event

to_event fut is an event that occurs once an infinitesimal amount of time after fut does. This means at earliest at the next JavaScript event loop cycle, even if fut is already determined.

val of_event : 'a Note.event -> 'a Fut.t option

of_event e is a future that determines an infinitesimal amount of time after the current or next occurence of e. This means at earliest at the next JavaScript event loop cycle, even if e occurs now. e is observed only once by the future. None is returned if e never occurs.