Module Ui.Float_selector

Floating point value selector.

The value is selected by a slider. See the styling information.

Selector

type t

val v : ?class':Jstr.t -> ?enabled:bool Note.signal -> ?min:float Note.signal -> ?max:float Note.signal -> ?step:float option Note.signal -> float Note.signal -> t

val action : t -> float Note.event

action b is an event that occurs when a new float is selected.

val enabled : t -> bool Note.signal

enabled b is a signal that is true iff the selector is enabled.

val el : t -> Brr.El.t

Styling