Module Audio.Node

Audio nodes.

Nodes

module Channel_count_mode : sig ... end

Channel count mode enumeration.

module Channel_interpretation : sig ... end

Channel intepretation enumeration.

type t

The type for AudioNode objects.

type node = t

See t.

type context

See Context.Base.t.

val as_target : t -> Brr.Ev.target

as_target n is n as an event target.

val context : t -> context

context n is the audio context of n.

val input_count : t -> int

input_count n is the number of inputs of n.

val output_count : t -> int

output_count n is the number of outputs of n.

val channel_count : t -> int

channel_count n is the number of channels used by n.

val set_channel_count : t -> int -> unit

set_channel_count n c sets channel_count of n to c.

val channel_count_mode : t -> Channel_count_mode.t

channel_count_mode n is the channel count mode of n.

val set_channel_count_mode : t -> Channel_count_mode.t -> unit

set_channel_count_mode n m sets the channel_count_mode of n to m.

val channel_interpretation : t -> Channel_interpretation.t

channel_interpretation n is the channel interpretation of n.

val set_channel_interpretation : t -> Channel_interpretation.t -> unit

set_channel_interpretation n i sets the channel_interpretation of n to i.

val connect_node : ?output:int -> ?input:int -> t -> dst:t -> unit

connect_node n ~output ~input ~dst connects n to dst.

val connect_param : ?output:int -> t -> dst:Param.t -> unit

connect_param n ~output ~dst connects n to dst.

val disconnect : t -> unit

disconnect n disconnects all outgoing connections.

val disconnect_node : ?output:int -> ?input:int -> t -> dst:t -> unit

disconnect_node n ~dst disconnects n from dst.

val disconnect_param : ?output:int -> t -> dst:Param.t -> unit

disconnect_param n ~dst disconnects n from dst.

Node types

module Analyser : sig ... end

Analyser nodes.

module Biquad_filter : sig ... end

Biquad filter nodes.

module Buffer_source : sig ... end

Buffer source nodes.

module Channel_merger : sig ... end

Channel merger nodes.

module Channel_splitter : sig ... end

Channel splitter nodes.

module Constant_source : sig ... end

Constant source nodes.

module Convolver : sig ... end

Convolver node.

module Delay : sig ... end

Delay node.

module Destination : sig ... end

Destination nodes.

module Dynamics_compressor : sig ... end

Dyamics compressor nodes.

module Gain : sig ... end

Gain nodes.

module Iir_filter : sig ... end

IIR filter nodes.

module Media_element_source : sig ... end

Media element source nodes.

module Media_stream_destination : sig ... end

Media stream destination nodes.

module Media_stream_source : sig ... end

Media stream source nodes.

module Media_stream_track_source : sig ... end

Media stream track source nodes.

module Oscillator : sig ... end

Oscillator nodes.

module Panner : sig ... end

Panner nodes.

module Stereo_panner : sig ... end

Stereo panner nodes.

module Wave_shaper : sig ... end

Wave shaper nodes.

module Worklet : sig ... end

Worklet nodes.