Module Node.Buffer_source

Buffer source nodes.

type opts

The type for AudioBufferSourceOptions.

val opts : ?buffer:Buffer.t -> ?detune:float -> ?loop:bool -> ?loop_start:float -> ?loop_end:float -> ?playback_rate:float -> unit -> opts

opts () are buffer source node options with given parameters.

type t

The type for AudioBufferSourceNode objects.

val create : ?opts:opts -> context -> t

create ~opts c creates a buffer source node.

val as_node : t -> node

as_node n is n as an audio node.

val buffer : t -> Buffer.t option

buffer n is the buffer of n.

val set_buffer : t -> Buffer.t option -> unit

set_buffer n b sets the buffer of n to b.

val playback_rate : t -> Param.t

playback_rate n is the playback rate of n.

val detune : t -> Param.t

detune n is the detune parameter of n.

val loop : t -> bool

loop n is the loop parameter of n.

val set_loop : t -> bool -> unit

set_loop n b sets the loop parameter of n to b.

val loop_start : t -> float

loop_start n is the loop_start parameter of n.

val set_loop_start : t -> float -> unit

set_loop_start n v sets the loop_start parameter of n to v.

val loop_end : t -> float

loop_end n is the loop_end parameter of n.

val set_loop_end : t -> float -> unit

set_loop_end n v sets the loop_end parameter of n to v.

val start : ?time:float -> ?offset:float -> ?dur_s:float -> t -> unit

start n starts the source.

val stop : ?time:float -> t -> unit

stop n stops the source.