Module Node.Convolver

Convolver node.

type opts

The type for ConvolverOptions.

val opts : ?channel_count:int -> ?channel_count_mode:Channel_count_mode.t -> ?channel_interpretation:Channel_interpretation.t -> ?buffer:Buffer.t -> ?disable_normalization:bool -> unit -> opts

opts () are channel splitter node options with given parameters.

type t

The type for ConvlverNode objects.

val create : ?opts:opts -> context -> t

create ~opts c creates a channel splitter node.

val as_node : t -> node

as_node n is n as an audio node.

val buffer : t -> Buffer.t option

buffer n is the buffer of n.

val set_buffer : t -> Buffer.t option -> unit

set_buffer n b sets the buffer of n to b.

val normalize : t -> bool

normalize n is the normalization behaviour of n.

val set_normalize : t -> bool -> unit

set_normalize n b sets normalize of n to b.