Module Node.Media_element_source

Media element source nodes.

type opts

The type for MediaElementAudioSourceOptions.

val opts : el:Brr_io.Media.El.t -> unit -> opts

opts ~el () are media element source node options with given parameters.

type t

The type for MediaElementAudioSourceNode objects.

val create : context -> opts:opts -> t

create ~opts c creates a media element audio source node.

val as_node : t -> node

as_node n is n as an audio node.

val media_element : t -> Brr_io.Media.El.t

media_element n is the media element of n.