Module Node.Media_stream_source

Media stream source nodes.

type opts

The type for MediaStreamAudioSourceOptions.

val opts : stream:Brr_io.Media.Stream.t -> unit -> opts

opts ~stream () are media stream source node options with given parameters.

type t

The type for MediaStreamAudioSourceNode objects.

val create : context -> opts:opts -> t

create ~opts c creates a media stream audio source node.

val as_node : t -> node

as_node n is n as an audio node.

val media_stream : t -> Brr_io.Media.Stream.t

media_stream n is the media stream of n.