Module Node.Wave_shaper

Wave shaper nodes.

module Oversample : sig ... end

Oversample type enumeration.

type opts

The type for WaveShaperOptions.

val opts : ?channel_count:int -> ?channel_count_mode:Channel_count_mode.t -> ?channel_interpretation:Channel_interpretation.t -> ?curve:Brr.Tarray.float32 -> ?oversample:Oversample.t -> unit -> opts

opts () are wave shaper node options with given parameters.

type t

The type for WaveShaperNode objects.

val create : ?opts:opts -> context -> t

create ~opts c creates a wave shaper node.

val as_node : t -> node

as_node n is n as an audio node.

val curve : t -> Brr.Tarray.float32 option

curve n is the curve of n.

val set_curve : t -> Brr.Tarray.float32 option -> unit

curve n v sets the curve of n to v.

val oversample : t -> Oversample.t

oversample n is the oversample of n.

val set_oversample : t -> Oversample.t -> unit

oversample n v sets the oversample of n to v.