Module Audio.Timestamp

Audio timestamps

type t

The type for AudioTimestamp objects.

val context_time : t -> float

context_time t is the time of the sample frame being rendered by the output device.

val performance_time : t -> float

performance_time t is an estimation of the time when the sample frame was rendered by the output device.