Module Brr_webcrypto.Crypto

Crypto objects.

type t

The type for Crypto objects.

val crypto : t

crypto is the global crypto object.

val subtle : t -> Subtle_crypto.t

subtle c is the stuble crypto object of c.

val set_random_values : t -> ('a, 'b) Brr.Tarray.t -> unit

set_random_values a overwrites the elements of a with random numbers. The function raises if the array is larger than 65535 bytes.