Module Crypto_algo.Aes_ctr_params

AES encryption parameters.

type t

The type for AesCtrParams objects.

val v : ?name:Jstr.t -> counter:Brr.Tarray.Buffer.t -> length:int -> unit -> algo

v ~name ~counter ~length is a encryption parameter object with given properties. name defaults to "AES-CTR".

val of_algo : algo -> t

of_algo a is an unsafe conversion from a.

Properties

val name : t -> Jstr.t

name a is the algorithm to use.

val counter : t -> Brr.Tarray.Buffer.t

counter a is the initial value of the counter block.

val length : t -> int

length a is the number of bits in the counter block used for the counter.