Module Crypto_algo.Aes_key_gen_params

AES key generation paramaters

type t

The type for AesKeyGenParams objects.

val v : name:Jstr.t -> length:int -> unit -> algo

v ~name ~hash is an key generation parameter object with given properties.

val of_algo : algo -> t

of_algo a is an unsafe conversion from a.

Properties

val name : t -> Jstr.t

name a is the algorithm to use.

val length : t -> int

length a is the bit length of the key.