Module Crypto_algo.Ecdh_key_derive_params

ECDH key derivation parameters.

type t

The type for EcdhKeyDeriveParams objects.

val v : name:Jstr.t -> public:Crypto_key.t -> unit -> algo

v ~name ~public is a signature parameter object with given properties.

val of_algo : algo -> t

of_algo a is an unsafe conversion from a.

Properties

val name : t -> Jstr.t

name a is the algorithm to use.

val public : t -> Jv.t

public a is the public key of the other entity.