Module Crypto_algo.Ecdsa_params

ECDSA signing parameters.

type t

The type for EcdsaParams objects.

val v : name:Jstr.t -> hash:Jstr.t -> unit -> algo

v ~name ~hash is a signature parameter object with given properties.

val of_algo : algo -> t

of_algo a is an unsafe conversion from a.

Properties

val name : t -> Jstr.t

name a is the algorithm to use.

val hash : t -> Jstr.t

hash a hash is the name of the hash to use.