Module Crypto_algo.Pbkdf2_params

PBKFD2 key derivation parameters.

type t

The type for Pbkdf2Params objects.

val v : ?name:Jstr.t -> hash:Jstr.t -> salt:Brr.Tarray.Buffer.t -> iterations:int -> unit -> algo

v ~name ~hash ~salt ~iterations is key derivation parameters object with given properties. name defaults to "PBKDF2".

val of_algo : algo -> t

of_algo a is an unsafe conversion from a.

Properties

val name : t -> Jstr.t

name a is the algorithm to use.

val hash : t -> Jstr.t

hash a is the name of the digest function to use.

val salt : t -> Brr.Tarray.Buffer.t

salt a is the random salt to use.

val iterations : t -> int

iteration a is the number of iterations to use.