Module Crypto_algo.Rsa_oaep_params

RSA-OAEP parameters.

type t

The type for RsaOaepParams objects.

val v : ?name:Jstr.t -> ?label:Brr.Tarray.Buffer.t -> unit -> algo

v ~name ~label is an encryption parameter object with given properties. name defaults to "RSA-OAEP".

val of_algo : algo -> t

of_algo a is an unsafe conversion from a.

Properties

val name : t -> Jstr.t

name a is the algorithm to use.

val label : t -> Brr.Tarray.Buffer.t option

label a is the label bound to the ciphertext.