Module Gpu.Adapter

Adapters.

module Info : sig ... end

Adapter info

type t

The type for GPUAdapter objects.

val has_feature : t -> Feature_name.t -> bool

has_feature a n is true iff n is a features of a.

val limits : t -> Supported_limits.t

limits a are the limits of a.

val is_fallback_adapter : t -> bool

is_fallback_adapter a indicates if a is a fallback adapter.

val request_device : ?descriptor:Device.Descriptor.t -> t -> Device.t Fut.or_error

request_device a requests the device of a.

val request_adapter_info : t -> unmask_hints:Jstr.t list -> Info.t Fut.or_error

request_adapter_info a ~unmask_hints requests the adapter info of a.