Module Command.Encoder

Encoders.

Encoders

module Descriptor : sig ... end

Descriptors.

type t

The type for GPUCommandEncoder objects.

val label : t -> Jstr.t

label e is the label of e.

val finish : ?descr:Buffer.Descriptor.t -> t -> Buffer.t

finish e finishes recording commands on e.

val begin_render_pass : t -> Render_pass.Descriptor.t -> Render_pass.Encoder.t

begin_render_pass begins encoding a render pass.

val begin_compute_pass : t -> Compute_pass.Descriptor.t -> Compute_pass.Encoder.t

begin_compute_pass begins encoding a compute pass.

val clear_buffer : ?size:int -> ?offset:int -> t -> buffer -> unit

clear_buffer clears a buffer.

val write_timestamp : t -> Query.Set.t -> int -> unit

write_timetamp writes a timetamp.

val resolve_query_set : t -> Query.Set.t -> first:int -> count:int -> dst:buffer -> dst_offset:int -> unit

resolve_query_set copies a query set in a buffer.

Copy commands

val copy_buffer_to_buffer : t -> src:buffer -> src_offset:int -> dst:buffer -> dst_offset:int -> size:int -> unit

copy_buffer_to_buffer copies betwen two buffers.

val copy_buffer_to_texture : t -> src:Image.Copy_buffer.t -> dst:Image.Copy_texture.t -> size:Extent_3d.t -> unit

copy_buffer_to_texture copies betwen a buffer and a texture.

val copy_texture_to_buffer : t -> src:Image.Copy_texture.t -> dst:Image.Copy_buffer.t -> size:Extent_3d.t -> unit

copy_texture_to_buffer copies betwen a texture and a buffer.

val copy_texture_to_texture : t -> src:Image.Copy_texture.t -> dst:Image.Copy_texture.t -> size:Extent_3d.t -> unit

copy_texture_to_texture copies betwen a texture and a texture.

Debug commands

val push_debug_group : t -> Jstr.t -> unit

push_debug_group e l starts a debug group l on e.

val pop_debug_group : t -> unit

pop_debug_group e ends the debug group o n e.

val insert_debug_marker : t -> Jstr.t -> unit

insert_debug_marker e l marks a point l in e.