Module Compute_pass.Encoder

Pass encoders.

Encoders

type t

The type for GPUComputePassEncoder objects.

val label : t -> Jstr.t

label e is the label of e.

val end' : t -> unit

end' e ends the pass on e.

Setup commands

val set_pipeline : t -> Compute_pipeline.t -> unit

set_pipeline e p sets the pipeline of e to p.

val set_bind_group : ?dynamic_offsets:int list -> ?group:Bind_group.t -> t -> index:int -> unit

set_bind_group sets a bind group for subsequent commands.

val set_bind_group' : ?group:Bind_group.t -> t -> index:int -> dynamic_offsets:int array -> offsets_start:int -> offsets_length:int -> unit

set_bind_group' sets a bind group for subsequent commands.

Dispatch commands

val dispatch_workgroups : ?count_z:int -> ?count_y:int -> t -> count_x:int -> unit

dispatch_workgroups dispaches a grid of workgroups.

val dispatch_workgroups_indirect : t -> Buffer.t -> offset:int -> unit

dispatch_workgroups_indirect dispatches a grid of workgroups.

Debug commands

val push_debug_group : t -> Jstr.t -> unit

push_debug_group e l starts a debug group pass l on e.

val pop_debug_group : t -> unit

pop_debug_group ends the debug group pass on e.

val insert_debug_marker : t -> Jstr.t -> unit

insert_debug_marker e l marks a point l in e.