Module Render_pass.Encoder

Encoders.

Encoders

type t

The type for GPURenderPassEncoder objects.

val label : t -> Jstr.t

label e is the label of e.

val end' : t -> unit

end e ends the command sequence.

Setup commands

val set_pipeline : t -> Render_pipeline.t -> unit

set_pipline e p sets the pipeline of e to p for subsequent commands.

val set_viewport : t -> x:float -> y:float -> w:float -> h:float -> min_depth:float -> max_depth:float -> unit

set_viewport sets the viewport.

val set_scissor_rect : t -> x:int -> y:int -> w:int -> h:int -> unit

set_scissor_rect sets the scissor rectangle.

val set_blend_constant : t -> Color.t -> unit

set_blend_constant sets the blend color constant.

val set_stencil_reference : t -> int -> unit

set_stencil_reference sets the stencil reference.

val set_bind_group : ?dynamic_offsets:int list -> ?group:Bind_group.t -> t -> index:int -> unit

set_bind_group sets a bind group for subsequent commands.

val set_bind_group' : ?group:Bind_group.t -> t -> index:int -> dynamic_offsets:int array -> offsets_start:int -> offsets_length:int -> unit

set_bind_group sets a bind group for subsequent commands.

set_bind_group' sets a bind group for subsequent commands.

val set_index_buffer : ?offset:int -> ?size:int -> t -> Buffer.t -> format:Index_format.t -> unit

set_index_buffer sets the index data for subsequent commands.

val set_vertex_buffer : ?buffer:Buffer.t -> ?offset:int -> ?size:int -> t -> slot:int -> unit

set_index_buffer sets sets the vertex data for subsequent commands.

Bundle and draw commands

val execute_bundles : t -> Render_bundle.t list -> unit

execute_bundles e bs executes bundles bs on e.

val draw : ?first_instance:int -> ?first_vertex:int -> ?instance_count:int -> t -> vertex_count:int -> unit

draw draws primitives.

val draw_indexed : ?first_instance:int -> ?base_vertex:int -> ?first_index:int -> ?instance_count:int -> t -> index_count:int -> unit

draw_indexed draws indexed primitives.

val draw_indirect : t -> Buffer.t -> offset:int -> unit

draw_indirect draws primitives.

val draw_indexed_indirect : t -> Buffer.t -> offset:int -> unit

draw_indexed_indirect draws indexed primitives.

Debug commands

val push_debug_group : t -> Jstr.t -> unit

push_debug_group e l starts a debug group l on e.

val pop_debug_group : t -> unit

pop_debug_group e ends the debug group o n e.

val insert_debug_marker : t -> Jstr.t -> unit

insert_debug_marker e l marks a point l in e.