Module Service_worker.G

Service worker global properties and functions.

See also Brr.G and Brr_io.Fetch.caches.

val clients : Clients.t

clients are the clients associated with the service worker.

val registration : Registration.t

registration is the service worker registration.

val service_worker : t

service_worker is the service worker.

val skip_waiting : unit -> unit Fut.or_error

skip_waiting () forces the waiting service to become the active service worker.