Module Service_worker.Registration

Service registration objects.

type t

The type for ServiceWorkerRegistration objects.

val as_target : t -> Brr.Ev.target

as_target r is r as an event target.

val update : t -> unit Fut.or_error

update r attempts to update the service worker of r.

val unregister : t -> bool Fut.or_error

unregister r unregisters the service worker registration. This is false if no registration was false.

val show_notification : ?opts:Brr_io.Notification.opts -> t -> Jstr.t -> unit Fut.or_error

show_notification r title ~opts displays a notification with title title an options opts.

val get_notifications : ?tag:Jstr.t -> t -> Brr_io.Notification.t list Fut.or_error

get_notifications r ~tag are notifications created via r and tagged with tag (or all of them if unspecified).

Properties

val installing : t -> service_worker option

installing r is the installing service worker of r.

val waiting : t -> service_worker option

waiting r is the installed service worker of r.

val active : t -> service_worker option

active r is the active service worker of r.

val navigation_preload : t -> Navigation_preload_manager.t

navigation_preload r is the navigation preload manager of r.

val scope : t -> Jstr.t

scope r is the scope of r.

val update_via_cache : t -> Update_via_cache.t

update_via_cache r is the update via cache property of r.