Module Worker.G

Worker global functions

val import_scripts : Jstr.t list -> unit

import_scripts uris synchronously imports the given scripts in the worker.

val post : ?opts:Brr_io.Message.opts -> 'a -> unit

post ~opts v posts value v on the global object with options opts (the target_origin option is meaningless in this case).

val close : unit -> unit

close () closes this worker scope.