Module Jv.Bool

bool properties accessors.

val find : t -> prop -> bool option

find o p is Option.map to_bool (find o p). This is unsafe, only use if you know that if o defines p it is guaranteed to be a JavaScript boolean.

val get : t -> prop -> bool

get o p is to_bool (get o p). This is unsafe, only use if you know o has p and it is guaranteed to be a JavaScript boolean.

val set : t -> prop -> bool -> unit

set o p b is set o p (of_bool b).

val set_if_some : t -> prop -> bool option -> unit

set_if_some o p b is set_if_some o p (Option.map of_bool b).