Module Jv.Jarray

JavaScript arrays.

Arrays

type t = jv

The type for JavaScript arrays.

val create : int -> t

create n is an array of length n.

val length : t -> int

length a is the array length.

Accessors

val get : t -> int -> t

get a i is the value of array a at index i.

val set : t -> int -> t -> unit

set a i sets the value of array a at index i to v.