Module El.Prop

Element properties.

type 'a t

The type for element properties of type 'a.

val bool : Jstr.t -> bool t

bool n is a property named n accessed as a bool and returning false on undefined.

val int : Jstr.t -> int t

int n is a property named n accessed as an int and returning 0 on undefined.

val float : Jstr.t -> float t

float n is a property named n accessed as a float and returning 0. on undefined.

val jstr : Jstr.t -> Jstr.t t

jstr n is the property named n accessed as a Jstr.t and returning Jstr.empty on undefined.

Predefined properties

val checked : bool t

val height : int t

val id : Jstr.t t

val name : Jstr.t t

val title : Jstr.t t

val value : Jstr.t t

val width : int t