Module Data_transfer.Item_list

DataTransferItemList objects.

type t

The type for DataTransferItemList objects.

val length : t -> int

length l it the length of the list.

val add_jstr : t -> type':Jstr.t -> Jstr.t -> Item.t option

add_jstr l type' s adds s with MIME type type'. None is returned in case of error, the corresponding data item otherwise.

val add_file : t -> File.t -> Item.t option

add_file l f adds file s. None is returned in case of error, the corresponding data item otherwise.

val remove : t -> int -> unit

remove l i removes the ith item from the list.

val clear : t -> unit

clear l removes all elements from the list.

val item : t -> int -> Item.t

item l i is the ith item in the list.

val items : t -> Item.t list

items l are the items of list l.