Module Performance.Entry

Performance entry objects.

module Type : sig ... end

Entry type enum.

type t

The type for PerformanceEntry objects.

val name : t -> Jstr.t

name e is the name of e.

val type' : t -> Type.t

type' e is the type of e.

val start_time : t -> float

start_time e is the start time of e.

val end_time : t -> float

end_time e is the end time of e.

val duration : t -> float

duration e is the duration of duration of e.

val to_json : t -> Json.t

to_json e is a JSON representation of e.

Entry types

Always check the type' before coercing an entry type to its subtype. Objects of type PerformanceMark, PerformanceMeasure, PerformanceFrameTiming, and PerformancePaintTiming do not provide additional properties, so no modules are provided for them.

type entry = t

See t.

module Resource_timing : sig ... end

Resource timing entries.

module Navigation_timing : sig ... end

Navigation timing entries.

val as_resource_timing : t -> Resource_timing.t

as_resource_timing e is e as a ressource timing entry.

val as_navigation_timing : t -> Navigation_timing.t

as_navigation_timing e is e as a navigation timing entry.