Module Clipboard.Item

Clipboard items.

module Presentation_style : sig ... end

Presentation style enum.

type opts

The type for ClipboardItemOptions.

val opts : ?presentation_style:Presentation_style.t -> unit -> opts

opts ~presentation_style () are options for clipboard item objects.

type t

The type for ClipboardItem objects.

val create : ?opts:opts -> (Jstr.t * Brr.Blob.t) list -> t

create ~opts data is clipboard item with MIME types and associated values data and options opts.

val presentation_style : t -> Presentation_style.t

presentation_style i is the presentation style of i.

val last_modified_ms : t -> int

last_modified_ms i is the last modified time in ms from the epoch of i.

val delayed : t -> bool

delayed i is the delayed property of i.

val types : t -> Jstr.t list

types i is the array of MIME types available for i.

val get_type : t -> Jstr.t -> Brr.Blob.t Fut.or_error

get_type i t is the blob object with MIME type t for item i.