Module Cache.Storage

Cache storage objects.

type cache = t

See t.

type t

The type for CacheStorage objects. See Brr_io.Fetch.caches to get one.

val match' : ?query_opts:query_opts -> t -> Request.t -> Response.t option Fut.or_error

match' s req is a stored response for req in s (if any).

val has : t -> Jstr.t -> bool Fut.or_error

has s n is true if n matches a cache name in s.

val open' : t -> Jstr.t -> cache Fut.or_error

open' s n opens the cache named n of s.

val delete : t -> Jstr.t -> bool Fut.or_error

delete s n deletes the cache named n from s. false is returned if n did not exist.

val keys : t -> Jstr.t list Fut.or_error

keys s are the cache names in s.