Module Message.Channel

Message channels.

See the Channel Messaging API.

type t

The type for MessageChannel objects.

val create : unit -> t

create () is a new channel.

val port1 : t -> Port.t

port c is the first port of c. The port attached to the context that created the channel.

val port2 : t -> Port.t

port2 c is the second port of c. The port attached to the context at the other end of the channel.