Module Message.Ev

Message events.

Message event object

type t

The type for MessageEvent and ExtendableMessageEvent objects.

val as_extendable : t -> Brr.Ev.Extendable.t

as_extendable e is e as an extendable event. Warning. only for ExtendableMessageEvents objects.

val data : t -> 'a

data e is the data send by the emitter. Warning. Unsafe, make sure to constrain the result value to the right type.

val origin : t -> Jstr.t

origin e is the origin of the message emitter.

val last_event_id : t -> Jstr.t

last_event_id e is a unique id for the event.

val source : t -> Jv.t option

source e is the message emitter.

val ports : t -> Port.t list

ports e is a list of ports associated with the channel the message is being send through (if applicable).

Events

val message : t Brr.Ev.type'

message is the message event.

val messageerror : t Brr.Ev.type'

messageerror is the messageerror event.