Module Storage.Ev

Storage event.

Storage event object

type storage_area = t

See Brr_io.Storage.t .

type t

The type for StorageEvent objects.

val key : t -> Jstr.t option

key e is the key of the item being changed.

val old_value : t -> Jstr.t option

old_value e is the old value of the key.

val new_value : t -> Jstr.t option

new_value e is the new value of the key.

val url : t -> Jstr.t

url e is the URL of the document whose storage item changed.

val storage_area : t -> storage_area option

storage_area e is the storage object.

Storage event

val storage : t Brr.Ev.type'

storage is the type for storage event fired on Brr.Windows on storage changes.