Module Audio.Listener

Audio listeners.

type t

The type for AudioListener objects.

val position_x : t -> Param.t

position_x l is the x position of l.

val position_y : t -> Param.t

position_y l is the y position of l.

val position_z : t -> Param.t

position_z l is the z position of l.

val forward_x : t -> Param.t

forward_x l is the x forward direction of l.

val forward_y : t -> Param.t

forward_y l is the y forward direction of l.

val forward_z : t -> Param.t

forward_z l is the z forward direction of l.

val up_x : t -> Param.t

up_x l is the x up direction of l.

val up_y : t -> Param.t

up_y l is the y up direction of l.

val up_z : t -> Param.t

up_z l is the z up direction of l.