Module Node.Panner

Panner nodes.

module Panning_model : sig ... end

Panning model type enumeration.

module Distance_model : sig ... end

Distance model type enumeration.

type opts

The type for PannerOptions.

val opts : ?channel_count:int -> ?channel_count_mode:Channel_count_mode.t -> ?channel_interpretation:Channel_interpretation.t -> ?panning_model:Panning_model.t -> ?distance_model:Distance_model.t -> ?position_x:float -> ?position_y:float -> ?position_z:float -> ?orientation_x:float -> ?orientation_y:float -> ?orientation_z:float -> ?ref_distance:float -> ?max_distance:float -> ?rolloff_factor:float -> ?cone_inner_angle:float -> ?cone_outer_angle:float -> ?cone_outer_gain:float -> unit -> opts

opts () are panner node options with given parameters.

type t

The type for PannerNode objects.

val create : ?opts:opts -> context -> t

create ~opts c creates a panner node.

val as_node : t -> node

as_node n is n as an audio node.

val panning_model : t -> Panning_model.t

panning_model n is the panning model of n.

val set_panning_model : t -> Panning_model.t -> unit

set_panning_model n v sets the panning_model of n to v.

val distance_model : t -> Distance_model.t

distance_model n is the distance model of n.

val set_distance_model : t -> Distance_model.t -> unit

set_distance_model n v sets the distance_model of n to v.

val position_x : t -> Param.t

position_x n is the x position of n.

val position_y : t -> Param.t

position_y n is the y position of n.

val position_z : t -> Param.t

position_z n is the z position of n.

val orientation_x : t -> Param.t

orientation_x n is the x orientation of n.

val orientation_y : t -> Param.t

orientation_y n is the y orientation of n.

val orientation_z : t -> Param.t

orientation_z n is the z orientation of n.

val ref_distance : t -> float

ref_distance n is the reference distance of n.

val set_ref_distance : t -> float -> unit

set_ref_distance n v sets the ref_distance of n to v.

val max_distance : t -> float

max_distance n is the maximal distance of n.

val set_max_distance : t -> float -> unit

set_max_distance n v sets the max_distance of n to v.

val cone_inner_angle : t -> float

cone_inner_angle n is the cone inner angle of n.

val set_cone_inner_angle : t -> float -> unit

set_cone_inner_angle n v sets the cone_inner_angle of n to v.

val cone_outer_angle : t -> float

cone_outer_angle n is the cone outer angle of n.

val set_cone_outer_angle : t -> float -> unit

set_cone_outer_angle n v sets the cone_outer_angle of n to v.

val cone_outer_gain : t -> float

cone_outer_gain n is the cone outer gain of n.

val set_cone_outer_gain : t -> float -> unit

set_cone_outer_gain n v sets the cone_outer_gain of n to v.