Module Service_worker.Navigation_preload_manager

Ressources preloading

type t

The type for NavigationPreload objects.

val enable : t -> unit Fut.or_error

enable p enables navigation preloading.

val disable : t -> unit Fut.or_error

disables p disables navigation preloading.

val set_header_value : t -> Jstr.t -> unit Fut.or_error

set_header_value p v sets the value of the header.

val get_state : t -> (bool * Jstr.t) Fut.or_error

get_state p indicates whether preload is enabled and the value of the header.